Split caching allocators off from CUDAService
