+++ CUDADataFormats/BeamSpot/BuildFile.xml
@@ -1,5 +1,6 @@
-<use name="FWCore/ServiceRegistry"/>
-<use name="HeterogeneousCore/CUDAServices"/>
+<use name="CUDADataFormats/Common"/>
+<use name="DataFormats/Common"/>
+<use name="HeterogeneousCore/CUDAUtilities"/>
 <use name="cuda-api-wrappers"/>
 <use name="rootcore"/>
