+++ CUDADataFormats/SiPixelCluster/BuildFile.xml
@@ -1,5 +1,6 @@
-<use name="FWCore/ServiceRegistry"/>
-<use name="HeterogeneousCore/CUDAServices"/>
+<use name="DataFormats/Common"/>
+<use name="CUDADataFormats/Common"/>
+<use name="HeterogeneousCore/CUDAUtilities"/>
 <use name="cuda-api-wrappers"/>
 <use name="rootcore"/>
