+++ CUDADataFormats/SiPixelDigi/BuildFile.xml
@@ -1,6 +1,7 @@
+<use name="CUDADataFormats/Common"/>
+<use name="DataFormats/Common"/>
 <use name="DataFormats/SiPixelRawData"/>
-<use name="FWCore/ServiceRegistry"/>
-<use name="HeterogeneousCore/CUDAServices"/>
+<use name="HeterogeneousCore/CUDAUtilities"/>
 <use name="cuda-api-wrappers"/>
 <use name="rootcore"/>
