+++ CUDADataFormats/TrackingRecHit/BuildFile.xml
@@ -1,8 +1,7 @@
 <use name="cuda-api-wrappers"/>
 <use name="rootcore"/>
-<use name="FWCore/ServiceRegistry"/>
-<use name="FWCore/ParameterSetReader"/>
-<use name="HeterogeneousCore/CUDAServices"/>
+<use name="CUDADataFormats/Common"/>
+<use name="DataFormats/Common"/>
 <use name="HeterogeneousCore/CUDAUtilities"/>
 <export>
     <lib name="1"/>
