+++ CUDADataFormats/TrackingRecHit/src/TrackingRecHit2DCUDA.cc
@@ -1,13 +1,12 @@
 #include "CUDADataFormats/TrackingRecHit/interface/TrackingRecHit2DCUDA.h"
-#include "FWCore/ServiceRegistry/interface/Service.h"
-#include "HeterogeneousCore/CUDAServices/interface/CUDAService.h"
 #include "HeterogeneousCore/CUDAUtilities/interface/copyAsync.h"
 #include "HeterogeneousCore/CUDAUtilities/interface/cudaCheck.h"
+#include "HeterogeneousCore/CUDAUtilities/interface/device_unique_ptr.h"
+#include "HeterogeneousCore/CUDAUtilities/interface/host_unique_ptr.h"
 
 template<>
 cudautils::host::unique_ptr<float[]> TrackingRecHit2DCUDA::localCoordToHostAsync(cuda::stream_t<> &stream) const {
-  edm::Service<CUDAService> cs;
-  auto ret = cs->make_host_unique<float[]>(4 * nHits(), stream);
+  auto ret = cudautils::make_host_unique<float[]>(4 * nHits(), stream);
   cudautils::copyAsync(ret, m_store32, 4 * nHits(), stream);
   return ret;
 }
@@ -15,8 +14,7 @@ cudautils::host::unique_ptr<float[]> TrackingRecHit2DCUDA::localCoordToHostAsync
 template<>
 cudautils::host::unique_ptr<uint32_t[]> TrackingRecHit2DCUDA::hitsModuleStartToHostAsync(
     cuda::stream_t<> &stream) const {
-  edm::Service<CUDAService> cs;
-  auto ret = cs->make_host_unique<uint32_t[]>(2001, stream);
+  auto ret = cudautils::make_host_unique<uint32_t[]>(2001, stream);
   cudaMemcpyAsync(ret.get(), m_hitsModuleStart, 4 * 2001, cudaMemcpyDefault, stream.id());
   return ret;
 }
