+++ CUDADataFormats/TrackingRecHit/test/TrackingRecHit2DCUDA_t.cpp
@@ -1,9 +1,4 @@
 #include "CUDADataFormats/TrackingRecHit/interface/TrackingRecHit2DCUDA.h"
-#include "FWCore/ParameterSet/interface/ConfigurationDescriptions.h"
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
-#include "FWCore/ParameterSetReader/interface/ParameterSetReader.h"
-#include "FWCore/PluginManager/interface/PluginManager.h"
-#include "FWCore/PluginManager/interface/standard.h"
 #include "HeterogeneousCore/CUDAUtilities/interface/copyAsync.h"
 #include "HeterogeneousCore/CUDAUtilities/interface/exitSansCUDADevices.h"
 
@@ -15,20 +10,6 @@ namespace testTrackingRecHit2D {
 
 int main() {
   exitSansCUDADevices();
-
-  edmplugin::PluginManager::configure(edmplugin::standard::config());
-
-  const std::string config{
-      R"_(import FWCore.ParameterSet.Config as cms
-process = cms.Process('Test')
-process.CUDAService = cms.Service('CUDAService')
-)_"};
-
-  std::unique_ptr<edm::ServiceRegistry::Operate> operate_;
-  std::unique_ptr<edm::ParameterSet> params;
-  edm::makeParameterSets(config, params);
-  edm::ServiceToken tempToken(edm::ServiceRegistry::createServicesFromConfig(std::move(params)));
-  operate_.reset(new edm::ServiceRegistry::Operate(tempToken));
 
   auto current_device = cuda::device::current::get();
   auto stream = current_device.create_stream(cuda::stream::implicitly_synchronizes_with_default_stream);
