+++ HeterogeneousCore/CUDATest/plugins/TestCUDAAnalyzerGPUKernel.cu
@@ -1,7 +1,7 @@
 #include "TestCUDAAnalyzerGPUKernel.h"
 
-#include "FWCore/ServiceRegistry/interface/Service.h"
-#include "HeterogeneousCore/CUDAServices/interface/CUDAService.h"
+#include "HeterogeneousCore/CUDAUtilities/interface/device_unique_ptr.h"
+#include "HeterogeneousCore/CUDAUtilities/interface/host_unique_ptr.h"
 #include "HeterogeneousCore/CUDAUtilities/interface/memsetAsync.h"
 #include "HeterogeneousCore/CUDAUtilities/interface/copyAsync.h"
 
@@ -25,8 +25,7 @@ namespace {
 }
 
 TestCUDAAnalyzerGPUKernel::TestCUDAAnalyzerGPUKernel(cuda::stream_t<>& stream) {
-  edm::Service<CUDAService> cs;
-  sum_ = cs->make_device_unique<float[]>(NUM_VALUES, stream);
+  sum_ = cudautils::make_device_unique<float[]>(NUM_VALUES, stream);
   cudautils::memsetAsync(sum_, 0, NUM_VALUES, stream);
   // better to synchronize since there is no guarantee that the stream
   // of analyzeAsync() would be otherise synchronized with this one
@@ -38,9 +37,8 @@ void TestCUDAAnalyzerGPUKernel::analyzeAsync(const float *d_input, cuda::stream_
 }
 
 float TestCUDAAnalyzerGPUKernel::value(cuda::stream_t<>& stream) const {
-  edm::Service<CUDAService> cs;
-  auto accumulator = cs->make_device_unique<float>(stream);
-  auto h_accumulator = cs->make_host_unique<float>(stream);
+  auto accumulator = cudautils::make_device_unique<float>(stream);
+  auto h_accumulator = cudautils::make_host_unique<float>(stream);
   sum<<<1,1, 0, stream.id()>>>(sum_.get(), accumulator.get(), NUM_VALUES);
   cudautils::copyAsync(h_accumulator, accumulator, stream);
   // need to synchronize
