+++ HeterogeneousCore/CUDATest/plugins/TestCUDAProducerGPUKernel.cu
@@ -2,8 +2,8 @@
 
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
 #include "FWCore/Utilities/interface/Exception.h"
-#include "FWCore/ServiceRegistry/interface/Service.h"
-#include "HeterogeneousCore/CUDAServices/interface/CUDAService.h"
+#include "HeterogeneousCore/CUDAUtilities/interface/device_unique_ptr.h"
+#include "HeterogeneousCore/CUDAUtilities/interface/host_unique_ptr.h"
 
 namespace {
   template<typename T>
@@ -73,18 +73,16 @@ cudautils::device::unique_ptr<float[]> TestCUDAProducerGPUKernel::runAlgo(const 
     }
   }
 
-  edm::Service<CUDAService> cs;
-
-  auto h_a = cs->make_host_unique<float[]>(NUM_VALUES, stream);
-  auto h_b = cs->make_host_unique<float[]>(NUM_VALUES, stream);
+  auto h_a = cudautils::make_host_unique<float[]>(NUM_VALUES, stream);
+  auto h_b = cudautils::make_host_unique<float[]>(NUM_VALUES, stream);
 
   for (auto i=0; i<NUM_VALUES; i++) {
     h_a[i] = i;
     h_b[i] = i*i;
   }
 
-  auto d_a = cs->make_device_unique<float[]>(NUM_VALUES, stream);
-  auto d_b = cs->make_device_unique<float[]>(NUM_VALUES, stream);
+  auto d_a = cudautils::make_device_unique<float[]>(NUM_VALUES, stream);
+  auto d_b = cudautils::make_device_unique<float[]>(NUM_VALUES, stream);
 
   cuda::memory::async::copy(d_a.get(), h_a.get(), NUM_VALUES*sizeof(float), stream.id());
   cuda::memory::async::copy(d_b.get(), h_b.get(), NUM_VALUES*sizeof(float), stream.id());
@@ -92,14 +90,14 @@ cudautils::device::unique_ptr<float[]> TestCUDAProducerGPUKernel::runAlgo(const 
   int threadsPerBlock {32};
   int blocksPerGrid = (NUM_VALUES + threadsPerBlock - 1) / threadsPerBlock;
 
-  auto d_c = cs->make_device_unique<float[]>(NUM_VALUES, stream);
+  auto d_c = cudautils::make_device_unique<float[]>(NUM_VALUES, stream);
   auto current_device = cuda::device::current::get();
   edm::LogVerbatim("TestHeterogeneousEDProducerGPU") << "  " << label << " GPU launching kernels device " << current_device.id() << " CUDA stream " << stream.id();
   vectorAdd<<<blocksPerGrid, threadsPerBlock, 0, stream.id()>>>(d_a.get(), d_b.get(), d_c.get(), NUM_VALUES);
 
-  auto d_ma = cs->make_device_unique<float[]>(NUM_VALUES*NUM_VALUES, stream);
-  auto d_mb = cs->make_device_unique<float[]>(NUM_VALUES*NUM_VALUES, stream);
-  auto d_mc = cs->make_device_unique<float[]>(NUM_VALUES*NUM_VALUES, stream);
+  auto d_ma = cudautils::make_device_unique<float[]>(NUM_VALUES*NUM_VALUES, stream);
+  auto d_mb = cudautils::make_device_unique<float[]>(NUM_VALUES*NUM_VALUES, stream);
+  auto d_mc = cudautils::make_device_unique<float[]>(NUM_VALUES*NUM_VALUES, stream);
   dim3 threadsPerBlock3{NUM_VALUES, NUM_VALUES};
   dim3 blocksPerGrid3{1,1};
   if(NUM_VALUES*NUM_VALUES > 32) {
