+++ HeterogeneousCore/CUDAUtilities/src/CachingHostAllocator.h
@@ -1,5 +1,5 @@
-#ifndef HeterogenousCore_CUDAServices_src_CachingDeviceAllocator_h
-#define HeterogenousCore_CUDAServices_src_CachingDeviceAllocator_h
+#ifndef HeterogenousCore_CUDAUtilities_src_CachingHostAllocator_h
+#define HeterogenousCore_CUDAUtilities_src_CachingHostAllocator_h
 
 /******************************************************************************
  * Copyright (c) 2011, Duane Merrill.  All rights reserved.
@@ -30,12 +30,12 @@
  ******************************************************************************/
 
 /**
- * Forked to CMSSW by Matti Kortelainen
- */ 
+ * Modified to cache pinned host allocations by Matti Kortelainen
+ */
 
 /******************************************************************************
- * Simple caching allocator for device memory allocations. The allocator is
- * thread-safe and capable of managing device allocations on multiple devices.
+ * Simple caching allocator for pinned host memory allocations. The allocator is
+ * thread-safe.
  ******************************************************************************/
 
 #include <cmath>
@@ -56,25 +56,26 @@ namespace notcub {
 
 
 /******************************************************************************
- * CachingDeviceAllocator (host use)
+ * CachingHostAllocator (host use)
  ******************************************************************************/
 
 /**
- * \brief A simple caching allocator for device memory allocations.
+ * \brief A simple caching allocator pinned host memory allocations.
  *
  * \par Overview
- * The allocator is thread-safe and stream-safe and is capable of managing cached
- * device allocations on multiple devices.  It behaves as follows:
+ * The allocator is thread-safe.  It behaves as follows:
+ *
+ * I presume the CUDA stream-safeness is not useful as to read/write
+ * from/to the pinned host memory one needs to synchronize anyway. The
+ * difference wrt. device memory is that in the CPU all operations to
+ * the device memory are scheduled via the CUDA stream, while for the
+ * host memory one can perform operations directly.
  *
  * \par
- * - Allocations from the allocator are associated with an \p active_stream.  Once freed,
- *   the allocation becomes available immediately for reuse within the \p active_stream
- *   with which it was associated with during allocation, and it becomes available for
- *   reuse within other streams when all prior work submitted to \p active_stream has completed.
  * - Allocations are categorized and cached by bin size.  A new allocation request of
  *   a given size will only consider cached allocations within the corresponding bin.
  * - Bin limits progress geometrically in accordance with the growth factor
- *   \p bin_growth provided during construction.  Unused device allocations within
+ *   \p bin_growth provided during construction.  Unused host allocations within
  *   a larger bin cache are not reused for allocation requests that categorize to
  *   smaller bin sizes.
  * - Allocation requests below (\p bin_growth ^ \p min_bin) are rounded up to
@@ -82,12 +83,12 @@ namespace notcub {
  * - Allocations above (\p bin_growth ^ \p max_bin) are not rounded up to the nearest
  *   bin and are simply freed when they are deallocated instead of being returned
  *   to a bin-cache.
- * - %If the total storage of cached allocations on a given device will exceed
- *   \p max_cached_bytes, allocations for that device are simply freed when they are
+ * - %If the total storage of cached allocations  will exceed
+ *   \p max_cached_bytes, allocations are simply freed when they are
  *   deallocated instead of being returned to their bin-cache.
  *
  * \par
- * For example, the default-constructed CachingDeviceAllocator is configured with:
+ * For example, the default-constructed CachingHostAllocator is configured with:
  * - \p bin_growth          = 8
  * - \p min_bin             = 3
  * - \p max_bin             = 7
@@ -95,10 +96,10 @@ namespace notcub {
  *
  * \par
  * which delineates five bin-sizes: 512B, 4KB, 32KB, 256KB, and 2MB
- * and sets a maximum of 6,291,455 cached bytes per device
+ * and sets a maximum of 6,291,455 cached bytes
  *
  */
-struct CachingDeviceAllocator
+struct CachingHostAllocator
 {
 
     //---------------------------------------------------------------------
@@ -121,53 +122,47 @@ struct CachingDeviceAllocator
     //---------------------------------------------------------------------
 
     /**
-     * Descriptor for device memory allocations
+     * Descriptor for pinned host memory allocations
      */
     struct BlockDescriptor
     {
-        void*           d_ptr;              // Device pointer
+        void*           d_ptr;              // Host pointer
         size_t          bytes;              // Size of allocation in bytes
         unsigned int    bin;                // Bin enumeration
         int             device;             // device ordinal
         cudaStream_t    associated_stream;  // Associated associated_stream
         cudaEvent_t     ready_event;        // Signal when associated stream has run to the point at which this block was freed
 
-        // Constructor (suitable for searching maps for a specific block, given its pointer and device)
-        BlockDescriptor(void *d_ptr, int device) :
+        // Constructor (suitable for searching maps for a specific block, given its pointer)
+        BlockDescriptor(void *d_ptr) :
             d_ptr(d_ptr),
             bytes(0),
             bin(INVALID_BIN),
-            device(device),
+            device(INVALID_DEVICE_ORDINAL),
             associated_stream(nullptr),
             ready_event(nullptr)
         {}
 
-        // Constructor (suitable for searching maps for a range of suitable blocks, given a device)
-        BlockDescriptor(int device) :
+        // Constructor (suitable for searching maps for a range of suitable blocks)
+        BlockDescriptor() :
             d_ptr(nullptr),
             bytes(0),
             bin(INVALID_BIN),
-            device(device),
+            device(INVALID_DEVICE_ORDINAL),
             associated_stream(nullptr),
             ready_event(nullptr)
         {}
 
-        // Comparison functor for comparing device pointers
+        // Comparison functor for comparing host pointers
         static bool PtrCompare(const BlockDescriptor &a, const BlockDescriptor &b)
         {
-            if (a.device == b.device)
-                return (a.d_ptr < b.d_ptr);
-            else
-                return (a.device < b.device);
+            return (a.d_ptr < b.d_ptr);
         }
 
         // Comparison functor for comparing allocation sizes
         static bool SizeCompare(const BlockDescriptor &a, const BlockDescriptor &b)
         {
-            if (a.device == b.device)
-                return (a.bytes < b.bytes);
-            else
-                return (a.device < b.device);
+            return (a.bytes < b.bytes);
         }
     };
 
@@ -186,9 +181,6 @@ struct CachingDeviceAllocator
 
     /// Set type for live blocks (ordered by ptr)
     typedef std::multiset<BlockDescriptor, Compare> BusyBlocks;
-
-    /// Map type of device ordinals to the number of cached bytes cached by each device
-    typedef std::map<int, TotalBytes> GpuCachedBytes;
 
 
     //---------------------------------------------------------------------
@@ -255,14 +247,14 @@ struct CachingDeviceAllocator
 
     size_t          min_bin_bytes;      /// Minimum bin size
     size_t          max_bin_bytes;      /// Maximum bin size
-    size_t          max_cached_bytes;   /// Maximum aggregate cached bytes per device
+    size_t          max_cached_bytes;   /// Maximum aggregate cached bytes
 
     const bool      skip_cleanup;       /// Whether or not to skip a call to FreeAllCached() when destructor is called.  (The CUDA runtime may have already shut down for statically declared allocators)
     bool            debug;              /// Whether or not to print (de)allocation events to stdout
 
-    GpuCachedBytes  cached_bytes;       /// Map of device ordinal to aggregate cached bytes on that device
-    CachedBlocks    cached_blocks;      /// Set of cached device allocations available for reuse
-    BusyBlocks      live_blocks;        /// Set of live device allocations currently in use
+    TotalBytes      cached_bytes;       /// Aggregate cached bytes
+    CachedBlocks    cached_blocks;      /// Set of cached pinned host allocations available for reuse
+    BusyBlocks      live_blocks;        /// Set of live pinned host allocations currently in use
 
 #endif // DOXYGEN_SHOULD_SKIP_THIS
 
@@ -273,11 +265,11 @@ struct CachingDeviceAllocator
     /**
      * \brief Constructor.
      */
-    CachingDeviceAllocator(
+    CachingHostAllocator(
         unsigned int    bin_growth,                             ///< Geometric growth factor for bin-sizes
         unsigned int    min_bin             = 1,                ///< Minimum bin (default is bin_growth ^ 1)
         unsigned int    max_bin             = INVALID_BIN,      ///< Maximum bin (default is no max bin)
-        size_t          max_cached_bytes    = INVALID_SIZE,     ///< Maximum aggregate cached bytes per device (default is no limit)
+        size_t          max_cached_bytes    = INVALID_SIZE,     ///< Maximum aggregate cached bytes (default is no limit)
         bool            skip_cleanup        = false,            ///< Whether or not to skip a call to \p FreeAllCached() when the destructor is called (default is to deallocate)
         bool            debug               = false)            ///< Whether or not to print (de)allocation events to stdout (default is no stderr output)
     :
@@ -305,9 +297,9 @@ struct CachingDeviceAllocator
      * - \p max_cached_bytes    = (\p bin_growth ^ \p max_bin) * 3) - 1 = 6,291,455 bytes
      *
      * which delineates five bin-sizes: 512B, 4KB, 32KB, 256KB, and 2MB and
-     * sets a maximum of 6,291,455 cached bytes per device
+     * sets a maximum of 6,291,455 cached bytes
      */
-    CachingDeviceAllocator(
+    CachingHostAllocator(
         bool skip_cleanup = false,
         bool debug = false)
     :
@@ -325,12 +317,12 @@ struct CachingDeviceAllocator
 
 
     /**
-     * \brief Sets the limit on the number bytes this allocator is allowed to cache per device.
+     * \brief Sets the limit on the number bytes this allocator is allowed to cache
      *
      * Changing the ceiling of cached bytes does not cause any allocations (in-use or
      * cached-in-reserve) to be freed.  See \p FreeAllCached().
      */
-    cudaError_t SetMaxCachedBytes(
+    void SetMaxCachedBytes(
         size_t max_cached_bytes)
     {
         // Lock
@@ -342,37 +334,29 @@ struct CachingDeviceAllocator
 
         // Unlock
         mutex.Unlock();
-
-        return cudaSuccess;
     }
 
 
     /**
-     * \brief Provides a suitable allocation of device memory for the given size on the specified device.
+     * \brief Provides a suitable allocation of pinned host memory for the given size.
      *
-     * Once freed, the allocation becomes available immediately for reuse within the \p active_stream
-     * with which it was associated with during allocation, and it becomes available for reuse within other
-     * streams when all prior work submitted to \p active_stream has completed.
+     * Once freed, the allocation becomes available immediately for reuse.
      */
-    cudaError_t DeviceAllocate(
-        int             device,                     ///< [in] Device on which to place the allocation
+    cudaError_t HostAllocate(
         void            **d_ptr,                    ///< [out] Reference to pointer to the allocation
         size_t          bytes,                      ///< [in] Minimum number of bytes for the allocation
         cudaStream_t    active_stream = nullptr)    ///< [in] The stream to be associated with this allocation
     {
         *d_ptr                          = nullptr;
-        int entrypoint_device           = INVALID_DEVICE_ORDINAL;
+        int device                      = INVALID_DEVICE_ORDINAL;
         cudaError_t error               = cudaSuccess;
 
-        if (device == INVALID_DEVICE_ORDINAL)
-        {
-            if (CubDebug(error = cudaGetDevice(&entrypoint_device))) return error;
-            device = entrypoint_device;
-        }
+        if (CubDebug(error = cudaGetDevice(&device))) return error;
 
         // Create a block descriptor for the requested allocation
         bool found = false;
-        BlockDescriptor search_key(device);
+        BlockDescriptor search_key;
+        search_key.device = device;
         search_key.associated_stream = active_stream;
         NearestPowerOf(search_key.bin, search_key.bytes, bin_growth, bytes);
 
@@ -396,30 +380,35 @@ struct CachingDeviceAllocator
                 search_key.bytes    = min_bin_bytes;
             }
 
-            // Iterate through the range of cached blocks on the same device in the same bin
+            // Iterate through the range of cached blocks in the same bin
             CachedBlocks::iterator block_itr = cached_blocks.lower_bound(search_key);
             while ((block_itr != cached_blocks.end())
-                    && (block_itr->device == device)
                     && (block_itr->bin == search_key.bin))
             {
                 // To prevent races with reusing blocks returned by the host but still
-                // in use by the device, only consider cached blocks that are
-                // either (from the active stream) or (from an idle stream)
-                if ((active_stream == block_itr->associated_stream) ||
-                    (cudaEventQuery(block_itr->ready_event) != cudaErrorNotReady))
-                {
+                // in use for transfers, only consider cached blocks that are from an idle stream
+                if(cudaEventQuery(block_itr->ready_event) != cudaErrorNotReady) {
                     // Reuse existing cache block.  Insert into live blocks.
                     found = true;
                     search_key = *block_itr;
                     search_key.associated_stream = active_stream;
+                    if(search_key.device != device) {
+                      // If "associated" device changes, need to re-create the event on the right device
+                      if (CubDebug(error = cudaSetDevice(search_key.device))) return error;
+                      if (CubDebug(error = cudaEventDestroy(search_key.ready_event))) return error;
+                      if (CubDebug(error = cudaSetDevice(device))) return error;
+                      if (CubDebug(error = cudaEventCreateWithFlags(&search_key.ready_event, cudaEventDisableTiming))) return error;
+                      search_key.device = device;
+                    }
+
                     live_blocks.insert(search_key);
 
                     // Remove from free blocks
-                    cached_bytes[device].free -= search_key.bytes;
-                    cached_bytes[device].live += search_key.bytes;
+                    cached_bytes.free -= search_key.bytes;
+                    cached_bytes.live += search_key.bytes;
 
-                    if (debug) _CubLog("\tDevice %d reused cached block at %p (%lld bytes) for stream %lld, event %lld (previously associated with stream %lld, event %lld).\n",
-                        device, search_key.d_ptr, (long long) search_key.bytes, (long long) search_key.associated_stream, (long long) search_key.ready_event, (long long) block_itr->associated_stream, (long long) block_itr->ready_event);
+                    if (debug) _CubLog("\tHost reused cached block at %p (%lld bytes) for stream %lld, event %lld on device %lld (previously associated with stream %lld, event %lld).\n",
+                        search_key.d_ptr, (long long) search_key.bytes, (long long) search_key.associated_stream, (long long) search_key.ready_event, (long long) search_key.device, (long long) block_itr->associated_stream, (long long) block_itr->ready_event);
 
                     cached_blocks.erase(block_itr);
 
@@ -435,19 +424,13 @@ struct CachingDeviceAllocator
         // Allocate the block if necessary
         if (!found)
         {
-            // Set runtime's current device to specified device (entrypoint may not be set)
-            if (device != entrypoint_device)
-            {
-                if (CubDebug(error = cudaGetDevice(&entrypoint_device))) return error;
-                if (CubDebug(error = cudaSetDevice(device))) return error;
-            }
-
             // Attempt to allocate
-            if (CubDebug(error = cudaMalloc(&search_key.d_ptr, search_key.bytes)) == cudaErrorMemoryAllocation)
+            // TODO: eventually support allocation flags
+            if (CubDebug(error = cudaHostAlloc(&search_key.d_ptr, search_key.bytes, cudaHostAllocDefault)) == cudaErrorMemoryAllocation)
             {
                 // The allocation attempt failed: free all cached blocks on device and retry
-                if (debug) _CubLog("\tDevice %d failed to allocate %lld bytes for stream %lld, retrying after freeing cached allocations",
-                      device, (long long) search_key.bytes, (long long) search_key.associated_stream);
+                if (debug) _CubLog("\tHost failed to allocate %lld bytes for stream %lld on device %lld, retrying after freeing cached allocations",
+                      (long long) search_key.bytes, (long long) search_key.associated_stream, (long long) search_key.device);
 
                 error = cudaSuccess;    // Reset the error we will return
                 cudaGetLastError();     // Reset CUDART's error
@@ -455,25 +438,24 @@ struct CachingDeviceAllocator
                 // Lock
                 mutex.Lock();
 
-                // Iterate the range of free blocks on the same device
-                BlockDescriptor free_key(device);
-                CachedBlocks::iterator block_itr = cached_blocks.lower_bound(free_key);
+                // Iterate the range of free blocks
+                CachedBlocks::iterator block_itr = cached_blocks.begin();
 
-                while ((block_itr != cached_blocks.end()) && (block_itr->device == device))
+                while ((block_itr != cached_blocks.end()))
                 {
                     // No need to worry about synchronization with the device: cudaFree is
                     // blocking and will synchronize across all kernels executing
                     // on the current device
-
-                    // Free device memory and destroy stream event.
-                    if (CubDebug(error = cudaFree(block_itr->d_ptr))) break;
+  
+                    // Free pinned host memory.
+                    if (CubDebug(error = cudaFreeHost(block_itr->d_ptr))) break;
                     if (CubDebug(error = cudaEventDestroy(block_itr->ready_event))) break;
 
                     // Reduce balance and erase entry
-                    cached_bytes[device].free -= block_itr->bytes;
+                    cached_bytes.free -= block_itr->bytes;
 
-                    if (debug) _CubLog("\tDevice %d freed %lld bytes.\n\t\t  %lld available blocks cached (%lld bytes), %lld live blocks (%lld bytes) outstanding.\n",
-                        device, (long long) block_itr->bytes, (long long) cached_blocks.size(), (long long) cached_bytes[device].free, (long long) live_blocks.size(), (long long) cached_bytes[device].live);
+                    if (debug) _CubLog("\tHost freed %lld bytes.\n\t\t  %lld available blocks cached (%lld bytes), %lld live blocks (%lld bytes) outstanding.\n",
+                        (long long) block_itr->bytes, (long long) cached_blocks.size(), (long long) cached_bytes.free, (long long) live_blocks.size(), (long long) cached_bytes.live);
 
                     cached_blocks.erase(block_itr);
 
@@ -487,7 +469,7 @@ struct CachingDeviceAllocator
                 if (error) return error;
 
                 // Try to allocate again
-                if (CubDebug(error = cudaMalloc(&search_key.d_ptr, search_key.bytes))) return error;
+                if (CubDebug(error = cudaHostAlloc(&search_key.d_ptr, search_key.bytes, cudaHostAllocDefault))) return error;
             }
 
             // Create ready event
@@ -497,103 +479,68 @@ struct CachingDeviceAllocator
             // Insert into live blocks
             mutex.Lock();
             live_blocks.insert(search_key);
-            cached_bytes[device].live += search_key.bytes;
+            cached_bytes.live += search_key.bytes;
             mutex.Unlock();
 
-            if (debug) _CubLog("\tDevice %d allocated new device block at %p (%lld bytes associated with stream %lld, event %lld).\n",
-                      device, search_key.d_ptr, (long long) search_key.bytes, (long long) search_key.associated_stream, (long long) search_key.ready_event);
-
-            // Attempt to revert back to previous device if necessary
-            if ((entrypoint_device != INVALID_DEVICE_ORDINAL) && (entrypoint_device != device))
-            {
-                if (CubDebug(error = cudaSetDevice(entrypoint_device))) return error;
-            }
+            if (debug) _CubLog("\tHost allocated new host block at %p (%lld bytes associated with stream %lld, event %lld on device %lld).\n",
+                      search_key.d_ptr, (long long) search_key.bytes, (long long) search_key.associated_stream, (long long) search_key.ready_event, (long long) search_key.device);
         }
 
-        // Copy device pointer to output parameter
+        // Copy host pointer to output parameter
         *d_ptr = search_key.d_ptr;
 
         if (debug) _CubLog("\t\t%lld available blocks cached (%lld bytes), %lld live blocks outstanding(%lld bytes).\n",
-            (long long) cached_blocks.size(), (long long) cached_bytes[device].free, (long long) live_blocks.size(), (long long) cached_bytes[device].live);
+            (long long) cached_blocks.size(), (long long) cached_bytes.free, (long long) live_blocks.size(), (long long) cached_bytes.live);
 
         return error;
     }
 
 
     /**
-     * \brief Provides a suitable allocation of device memory for the given size on the current device.
+     * \brief Frees a live allocation of pinned host memory, returning it to the allocator.
      *
-     * Once freed, the allocation becomes available immediately for reuse within the \p active_stream
-     * with which it was associated with during allocation, and it becomes available for reuse within other
-     * streams when all prior work submitted to \p active_stream has completed.
+     * Once freed, the allocation becomes available immediately for reuse.
      */
-    cudaError_t DeviceAllocate(
-        void            **d_ptr,                    ///< [out] Reference to pointer to the allocation
-        size_t          bytes,                      ///< [in] Minimum number of bytes for the allocation
-        cudaStream_t    active_stream = nullptr)    ///< [in] The stream to be associated with this allocation
-    {
-        return DeviceAllocate(INVALID_DEVICE_ORDINAL, d_ptr, bytes, active_stream);
-    }
-
-
-    /**
-     * \brief Frees a live allocation of device memory on the specified device, returning it to the allocator.
-     *
-     * Once freed, the allocation becomes available immediately for reuse within the \p active_stream
-     * with which it was associated with during allocation, and it becomes available for reuse within other
-     * streams when all prior work submitted to \p active_stream has completed.
-     */
-    cudaError_t DeviceFree(
-        int             device,
+    cudaError_t HostFree(
         void*           d_ptr)
     {
         int entrypoint_device           = INVALID_DEVICE_ORDINAL;
         cudaError_t error               = cudaSuccess;
-
-        if (device == INVALID_DEVICE_ORDINAL)
-        {
-            if (CubDebug(error = cudaGetDevice(&entrypoint_device)))
-                return error;
-            device = entrypoint_device;
-        }
 
         // Lock
         mutex.Lock();
 
         // Find corresponding block descriptor
         bool recached = false;
-        BlockDescriptor search_key(d_ptr, device);
+        BlockDescriptor search_key(d_ptr);
         BusyBlocks::iterator block_itr = live_blocks.find(search_key);
         if (block_itr != live_blocks.end())
         {
             // Remove from live blocks
             search_key = *block_itr;
             live_blocks.erase(block_itr);
-            cached_bytes[device].live -= search_key.bytes;
+            cached_bytes.live -= search_key.bytes;
 
             // Keep the returned allocation if bin is valid and we won't exceed the max cached threshold
-            if ((search_key.bin != INVALID_BIN) && (cached_bytes[device].free + search_key.bytes <= max_cached_bytes))
+            if ((search_key.bin != INVALID_BIN) && (cached_bytes.free + search_key.bytes <= max_cached_bytes))
             {
                 // Insert returned allocation into free blocks
                 recached = true;
                 cached_blocks.insert(search_key);
-                cached_bytes[device].free += search_key.bytes;
+                cached_bytes.free += search_key.bytes;
 
-                if (debug) _CubLog("\tDevice %d returned %lld bytes at %p from associated stream %lld, event %lld.\n\t\t %lld available blocks cached (%lld bytes), %lld live blocks outstanding. (%lld bytes)\n",
-                    device, (long long) search_key.bytes, d_ptr, (long long) search_key.associated_stream, (long long) search_key.ready_event, (long long) cached_blocks.size(),
-                    (long long) cached_bytes[device].free, (long long) live_blocks.size(), (long long) cached_bytes[device].live);
+                if (debug) _CubLog("\tHost returned %lld bytes from associated stream %lld, event %lld on device %lld.\n\t\t %lld available blocks cached (%lld bytes), %lld live blocks outstanding. (%lld bytes)\n",
+                    (long long) search_key.bytes, (long long) search_key.associated_stream, (long long) search_key.ready_event, (long long) search_key.device, (long long) cached_blocks.size(),
+                    (long long) cached_bytes.free, (long long) live_blocks.size(), (long long) cached_bytes.live);
             }
         }
 
-        // First set to specified device (entrypoint may not be set)
-        if (device != entrypoint_device)
-        {
-            if (CubDebug(error = cudaGetDevice(&entrypoint_device))) return error;
-            if (CubDebug(error = cudaSetDevice(device))) return error;
+        if (CubDebug(error = cudaGetDevice(&entrypoint_device))) return error;
+        if (entrypoint_device != search_key.device) {
+            if (CubDebug(error = cudaSetDevice(search_key.device))) return error;
         }
 
-        if (recached)
-        {
+        if (recached) {
             // Insert the ready event in the associated stream (must have current device set properly)
             if (CubDebug(error = cudaEventRecord(search_key.ready_event, search_key.associated_stream))) return error;
         }
@@ -604,15 +551,15 @@ struct CachingDeviceAllocator
         if (!recached)
         {
             // Free the allocation from the runtime and cleanup the event.
-            if (CubDebug(error = cudaFree(d_ptr))) return error;
+            if (CubDebug(error = cudaFreeHost(d_ptr))) return error;
             if (CubDebug(error = cudaEventDestroy(search_key.ready_event))) return error;
 
-            if (debug) _CubLog("\tDevice %d freed %lld bytes at %p from associated stream %lld, event %lld.\n\t\t  %lld available blocks cached (%lld bytes), %lld live blocks (%lld bytes) outstanding.\n",
-                device, (long long) search_key.bytes, d_ptr, (long long) search_key.associated_stream, (long long) search_key.ready_event, (long long) cached_blocks.size(), (long long) cached_bytes[device].free, (long long) live_blocks.size(), (long long) cached_bytes[device].live);
+            if (debug) _CubLog("\tHost freed %lld bytes from associated stream %lld, event %lld on device %lld.\n\t\t  %lld available blocks cached (%lld bytes), %lld live blocks (%lld bytes) outstanding.\n",
+                (long long) search_key.bytes, (long long) search_key.associated_stream, (long long) search_key.ready_event, (long long) search_key.device, (long long) cached_blocks.size(), (long long) cached_bytes.free, (long long) live_blocks.size(), (long long) cached_bytes.live);
         }
 
         // Reset device
-        if ((entrypoint_device != INVALID_DEVICE_ORDINAL) && (entrypoint_device != device))
+        if ((entrypoint_device != INVALID_DEVICE_ORDINAL) && (entrypoint_device != search_key.device))
         {
             if (CubDebug(error = cudaSetDevice(entrypoint_device))) return error;
         }
@@ -622,21 +569,7 @@ struct CachingDeviceAllocator
 
 
     /**
-     * \brief Frees a live allocation of device memory on the current device, returning it to the allocator.
-     *
-     * Once freed, the allocation becomes available immediately for reuse within the \p active_stream
-     * with which it was associated with during allocation, and it becomes available for reuse within other
-     * streams when all prior work submitted to \p active_stream has completed.
-     */
-    cudaError_t DeviceFree(
-        void*           d_ptr)
-    {
-        return DeviceFree(INVALID_DEVICE_ORDINAL, d_ptr);
-    }
-
-
-    /**
-     * \brief Frees all cached device allocations on all devices
+     * \brief Frees all cached pinned host allocations
      */
     cudaError_t FreeAllCached()
     {
@@ -664,15 +597,15 @@ struct CachingDeviceAllocator
                 current_device = begin->device;
             }
 
-            // Free device memory
-            if (CubDebug(error = cudaFree(begin->d_ptr))) break;
+            // Free host memory
+            if (CubDebug(error = cudaFreeHost(begin->d_ptr))) break;
             if (CubDebug(error = cudaEventDestroy(begin->ready_event))) break;
 
             // Reduce balance and erase entry
-            cached_bytes[current_device].free -= begin->bytes;
+            cached_bytes.free -= begin->bytes;
 
-            if (debug) _CubLog("\tDevice %d freed %lld bytes.\n\t\t  %lld available blocks cached (%lld bytes), %lld live blocks (%lld bytes) outstanding.\n",
-                current_device, (long long) begin->bytes, (long long) cached_blocks.size(), (long long) cached_bytes[current_device].free, (long long) live_blocks.size(), (long long) cached_bytes[current_device].live);
+            if (debug) _CubLog("\tHost freed %lld bytes.\n\t\t  %lld available blocks cached (%lld bytes), %lld live blocks (%lld bytes) outstanding.\n",
+                (long long) begin->bytes, (long long) cached_blocks.size(), (long long) cached_bytes.free, (long long) live_blocks.size(), (long long) cached_bytes.live);
 
             cached_blocks.erase(begin);
         }
@@ -692,7 +625,7 @@ struct CachingDeviceAllocator
     /**
      * \brief Destructor
      */
-    virtual ~CachingDeviceAllocator()
+    ~CachingHostAllocator()
     {
         if (!skip_cleanup)
             FreeAllCached();
